backend/requirements.txt
amqp==5.0.5
attrs==20.3.0
autopep8==1.5.5
billiard==3.6.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.2
dj-database-url==0.5.0
Django==2.2.17
django-cors-headers==3.7.0
django-filter==2.4.0
django-jsonfield-backport==1.0.3
django-timezone-field==4.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.13.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
kombu==5.0.2
mccabe==0.6.1
ntlm-auth==1.5.0
prompt-toolkit==3.0.17
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-ntlm2==6.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
sentry-sdk==1.0.0
